update some id to 8bit and add some new columns

+++ database/migrations/2018_05_12_070922_create_ts_schedules_table.php
@@ -17,12 +17,11 @@ class CreateTsSchedulesTable extends Migration
         Schema::create('ts_schedules', function (Blueprint $table) {
             
             // Keys
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->unsignedInteger('client_id');
             $table->unsignedInteger('creator_id');
             $table->unsignedInteger('user_id');
-            $table->unsignedInteger('job_id')
-                ->nullable();     
+            $table->unsignedBigInteger('job_id')->nullable();     
             
             // Attributes
             $table->string('type');
@@ -30,14 +29,13 @@ class CreateTsSchedulesTable extends Migration
             $table->string('status');                      
             $table->dateTime('start');
             $table->dateTime('end');
-            $table->boolean('billable')
-                ->default(false);
-            $table->boolean('offsite')
-                ->default(false);
-            $table->time('break_length')
-                ->default('00:00');
-            $table->json('billing')
-                ->nullable();
+            $table->boolean('billable')->default(false);
+            $table->boolean('offsite')->default(false);
+            $table->time('break_length')->default('00:00');
+
+            // json
+            $table->json('billing')->nullable();
+            
             $table->softDeletes();
             $table->timestamps();
         });
